save system, shop items display

+++ Assets/Scripts/UpdateScoreUI.cs
@@ -5,7 +5,7 @@ public class UpdateScoreUI : MonoBehaviour
 {
     [SerializeField] private TextMeshProUGUI _counterText;
 
-    public void UpdateUI(int amount)
+    public void UpdateUI(ulong amount)
     {
         _counterText.text = $"${amount}";
     }
